Brings in the ID-photo-to-Registrar feature and the schema.sql/live-DB
sync session (which fixed two SQL type mismatches: user_security_answers.slot
and users.failed_security_attempts were TINYINT but their JPA fields are
Integer). Resolved conflicts in 5 memory-bank docs only — all application
code, including both branches' edits to the same migration file, merged
automatically with zero conflicts.

- changeLog.md, decisions.md, progress.md: kept both sides' entries
  (independent, non-contradictory dated log entries).
- activeContext.md: rewrote the singular "Current Phase" statement (both
  branches had written a different one) and relabeled both prior "Latest
  Session" write-ups as "Previous Session".
- testing.md: kept both sides' verification sections; the single
  "Latest full-suite result" line was resolved by actually running the full
  suite against the merged code rather than picking either pre-merge number.

Verified after resolving: compiles clean, 374 tests / 0 failures / 0 errors.

+++ src/main/sql/schema.sql
@@ -143,7 +143,7 @@ CREATE TABLE IF NOT EXISTS users (
     -- Security-question lockout state — deliberately separate from `enabled`
     -- (admin deactivate/re-enable). See migrations/2026-09-19-security-questions.sql.
     security_locked TINYINT(1) NOT NULL DEFAULT 0,
-    failed_security_attempts TINYINT NOT NULL DEFAULT 0,
+    failed_security_attempts INT NOT NULL DEFAULT 0,
     security_lockout_started_at DATETIME NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
 
@@ -175,7 +175,7 @@ CREATE TABLE IF NOT EXISTS user_security_answers (
     question_id INT NULL,
     custom_question VARCHAR(255) NULL,
     answer_hash VARCHAR(255) NOT NULL,
-    slot TINYINT NOT NULL,
+    slot INT NOT NULL,
     created_at DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (user_id) REFERENCES users (user_id) ON DELETE CASCADE,
     FOREIGN KEY (question_id) REFERENCES security_questions (question_id),
